+++ lsteamclient/cppISteamNetworkingUtils_SteamNetworkingUtils004.h
@@ -2,7 +2,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils004_AllocateMessage_params
 {
     void *linux_side;
     winSteamNetworkingMessage_t_158 *_ret;
-    int cbAllocateBuffer;
+    int32_t cbAllocateBuffer;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_AllocateMessage( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_AllocateMessage_params *params );
 
@@ -25,7 +25,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetLocalPingLocatio
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_EstimatePingTimeBetweenTwoLocations_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *location1;
     const SteamNetworkPingLocation_t *location2;
 };
@@ -34,7 +34,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_EstimatePingTimeBet
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_EstimatePingTimeFromLocalHost_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *remoteLocation;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_EstimatePingTimeFromLocalHost( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_EstimatePingTimeFromLocalHost_params *params );
@@ -44,7 +44,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils004_ConvertPingLocationToStr
     void *linux_side;
     const SteamNetworkPingLocation_t *location;
     char *pszBuf;
-    int cchBufSize;
+    int32_t cchBufSize;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_ConvertPingLocationToString( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_ConvertPingLocationToString_params *params );
 
@@ -68,7 +68,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_CheckPingDataUpToDa
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPingToDataCenter_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
     SteamNetworkingPOPID *pViaRelayPoP;
 };
@@ -77,7 +77,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPingToDataCenter
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetDirectPingToPOP_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetDirectPingToPOP( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetDirectPingToPOP_params *params );
@@ -85,16 +85,16 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetDirectPingToPOP(
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPCount_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPCount( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPCount_params *params );
 
 struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPList_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID *list;
-    int nListSz;
+    int32_t nListSz;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPList( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetPOPList_params *params );
 
@@ -117,7 +117,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetIPv4FakeIPType_params
 {
     void *linux_side;
     uint32_t _ret;
-    uint32 nIPv4;
+    uint32_t nIPv4;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils004_GetIPv4FakeIPType( struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetIPv4FakeIPType_params *params );
 
@@ -136,7 +136,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils004_SetConfigValue_params
     bool _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t eDataType;
     const void *pArg;
 };
@@ -148,7 +148,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils004_GetConfigValue_params
     uint32_t _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t *pOutDataType;
     void *pResult;
     size_t *cbResult;
